Revert "Refactor services to account for singal participant per tourney"
This reverts commit 53b7f613da8a6d2d65f695876a763fedb2d8ba4b.

+++ src/main/java/xyz/zlatanov/ravenscore/web/model/toursummary/TournamentSummaryModel.java
@@ -11,7 +11,7 @@ public class TournamentSummaryModel {
 
 	private String	id;
 	private String	name;
-	private Long	numberOfParticipants;
+	private Integer	numberOfParticipants;
 	private String	statusLabel;
 	private String	statusDate;
 	private boolean	pinned;
